# app/blueprints/service_tickets/schemas.py
from marshmallow import Schema, fields
from app.extensions import ma
from app.models import Service_tickets

class ServiceTicketSchema(ma.SQLAlchemyAutoSchema):
    class Meta:
        model = Service_tickets
        load_instance = False
        include_fk = True

service_ticket_schema = ServiceTicketSchema(partial=True)
service_tickets_schema = ServiceTicketSchema(many=True)


class EditServiceTicketSchema(Schema):
    add_mechanic_ids = fields.List(fields.Int(), load_default=[])
    remove_mechanic_ids = fields.List(fields.Int(), load_default=[])

edit_service_ticket_schema = EditServiceTicketSchema()