# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  version: "1.0.0"
  description: "API for managing customers, mechanics, service tickets and inventory."
host: "127.0.0.1:5000"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/login:
    post:
      tags:
        - customers
      summary: "Customer login"
      parameters:
        - in: body
          name: credentials
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: Login successful
          schema:
            $ref: "#/definitions/LoginResponse"
  /customers:
    post:
      tags:
        - customers
      summary: Create a new customer
      parameters:
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerInput"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/Customer"
    get:
      tags:
        - customers
      summary: List customers (paginated)
      parameters:
        - in: query
          name: page
          type: integer
        - in: query
          name: per_page
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              customers:
                type: array
                items:
                  $ref: "#/definitions/Customer"
              total:
                type: integer
              page:
                type: integer
              pages:
                type: integer
  /customers/{customer_id}:
    get:
      tags:
        - customers
      summary: Get a single customer
      parameters:
        - in: path
          name: customer_id
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Customer"
        404:
          description: Not Found
    put:
      tags:
        - customers
      summary: Update a customer (authenticated)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customer_id
          required: true
          type: integer
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerInput"
      responses:
        200:
          description: Updated
          schema:
            $ref: "#/definitions/Customer"
        403:
          description: Unauthorized
    delete:
      tags:
        - customers
      summary: Delete a customer (authenticated)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customer_id
          required: true
          type: integer
      responses:
        200:
          description: Deleted
          schema:
            type: object
            properties:
              message:
                type: string
  /customers/my-tickets:
    get:
      tags:
        - customers
      summary: Get tickets for authenticated customer
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/ServiceTicket"
...
definitions:
  LoginCredentials:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
  LoginResponse:
    type: object
    properties:
      status:
        type: string
      token:
        type: string
  Customer:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
  CustomerInput:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      password:
        type: string
    required:
      - name
      - email
      - phone
      - password
  Mechanic:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number
...
  Inventory:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
...
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
      customer_id:
        type: integer
      service_date:
        type: string
      issue_description:
        type: string
      vin:
        type: string
      mechanics:
        type: array
        items:
          $ref: "#/definitions/Mechanic"
      parts:
        type: array
        items:
          $ref: "#/definitions/Inventory"
...